requirements.txt
wheel
boto3==1.17.86
botocore==1.20.86
click==8.0.1
Flask==2.0.1
hurry.filesize==0.9
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
peewee==3.14.4
python-dateutil==2.8.1
python-decouple==3.4
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
gunicorn
requests
mysql-connector-python
flask_login
flask_wtf
flask_cors
flask_caching
mysql-connector-python
yaml
pyyaml